# OTT-Platform-Web-Application/Backend/docs/api_docs.yaml
...
openapi: 3.0.3
info:
  title: Dhaakad API
  description: Dhaakad API documentation (Might be buggy)
  version: 0.0.5
  contact:
    email: vishal.rustagi@metaorangedigital.com
servers:
  - url: 'https://api.dhaakadcinema.com/api/v1'
...
paths:
...
  /users/expireCoupon/{id}:
    patch:
      tags:
        - users
      summary: expire coupon
      operationId: expireCoupon
      parameters:
        - name: id
          in: path
          description: user id
          required: true
          schema:
            type: string
            example: 'abc123'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                coupons:
                  type: array
                  items:
                    type: string
                  example: ['abc', 'xyz']
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "admin||user updated successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/expireReward/{id}:
    patch:
      tags:
        - users
      summary: expire reward
      operationId: expireReward
      parameters:
        - name: id
          in: path
          description: user id
          required: true
          schema:
            type: string
            example: 'abc123'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "admin||user updated successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/getUserByRefCode/{id}:
    get:
      tags:
        - users
      summary: get User By Reference Code
      operationId: getUserByRefCode
      parameters:
        - name: id
          in: path
          description: user id
          required: true
          schema:
            type: string
            example: 'abc123'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "User fetched successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/addReward/{id}:
    patch:
      tags:
        - users
      summary: add reward
      operationId: addReward
      parameters:
        - name: id
          in: path
          description: user id
          required: true
          schema:
            type: string
            example: 'abc123'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Reward purchased successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users:
    get:
      tags:
        - users
      summary: get all Users
      operationId: getAllUsers
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Users fetched successfully"
                  results:
                    type: integer
                    example: 10
                  data:
                    type: object
                    properties:
                      user:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - users
      summary: add user
      description: add user
      operationId: addUser
      requestBody:
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/User'
                - type: object
                  properties:
                    photo:
                      type: string
                      format: binary
      responses:
        201:
          description: account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "admin||user created successfully"
                  token:
                    type: string
                    example: 'asdaSJKASNDKJNJKQNWKJEQOIE8QE92EQWJKMMMMMMMMMMMMMMMMMADN'
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'abc123'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "admin||user fetched successfully"
                  data:
                    type: object
                    properties:
                      user:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - users
      summary: update user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'abc123'
      requestBody:
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/User'
                - type: object
                  properties:
                    photo:
                      type: string
                      format: binary
      responses:
        200:
          description: user updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "admin||user updated successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - users
      summary: delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'abc123'
      responses:
        200:
          description: user updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Users deleted successfully"
                  data:
                    type: string
                    example: null

        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /rewards:
    get:
      tags:
        - rewards
      summary: Get all rewards
      operationId: getAllRewards
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Rewards fetched successfully"
                  results:
                    type: integer
                    example: 10
                  data:
                    type: object
                    properties:
                      reward:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reward'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - rewards
      summary: Add reward
      operationId: addReward
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 'New Subscription'
                subtitle:
                  type: string
                  example: 'English'
                couponCode:
                  type: string
                  example: 'asd'
                amount:
                  type: integer
                  example: 100
                description:
                  type: string
                  example: 'asdasd'
                expireDate:
                  type: string
                  format: date
                  example: '2024-08-12'
                userLimit:
                  type: string
                  example: 4
                status:
                  type: string
                  example: 'Active||Inactive'
                movieId:
                  type: string
                  example: 'asdasd'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Rewards created successfully"
                  data:
                    type: object
                    properties:
                      reward:
                        $ref: '#/components/schemas/Reward'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'asdasd'
        firstName:
          type: string
          example: 'Sandy'
        lastName:
          type: string
          example: 'Joseph'
        userName:
          type: string
          example: 'Sandy'
        email:
          type: string
          example: 'sandy@gmail.com'
        referralCode:
          type: string
          example: 'test'
        idReferralCode:
          type: boolean
          example: true
        rewards:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: 'Referral Reward'
              amount:
                type: number
                example: 10
        coupons:
          type: array
          items:
            type: string
          example: ['abc','xyz']
        otp:
          type: number
          example: '123456'
        expiryotp:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        password:
          type: string
          example: 'abcd'
        passwordConfirm:
          type: string
          example: 'abcd'
        mobile:
          type: 'string'
          example: '+91-1234567891'
        gender:
          type: string
          example: 'male||female'
        dateOfBirth:
          type: string
          format: date
          example: '1991-07-30'
        city:
          type: string
          example: 'Delhi'
        releasedBookedMovies:
          type: array
          items:
            type: string
          example: []
        preBookedMovies:
          type: array
          items:
            type: string
          example: []
        country:
          type: string
          example: 'India'
        zipCode:
          type: string
          example: '123456'
        state:
          type: string
          example: 'Delhi'
        status:
          type: string
          example: 'Active||Inactive'
        photo:
          type: string
          example: 'https://123.png'
        address:
          type: string
          example: '#123 5th cross Delhi'
        userType:
          type: string
          example: 'admin||user'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        firebaseId:
          type: string
          example: 'asdasd'
        passwordChangedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        passwordResetToken:
          type: string
          example: 'sadhq7ygwiueqwhe2t6783qeidasnsbhgdas'
        passwordResetExpires:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    Reward:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'pqr123'
        title:
          type: string
          example: 'New Subscription'
        subtitle:
          type: string
          example: 'English'
        couponCode:
          type: string
          example: 'asd'
        amount:
          type: integer
          example: 100
        description:
          type: string
          example: 'asdasd'
        expireDate:
          type: string
          format: date
          example: '2024-08-12'
        userLimit:
          type: string
          example: 4
        status:
          type: string
          example: 'Active||Inactive'
        movieId:
          type: string
          example: 'asdasd'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: 'Something went wrong'
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes